backend/requirements.txt
# /mnt/c/Users/asadi/Desktop/ChatBot/ChatBot_Reop/IndustrialProject/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
llama-index-core>=0.10.0
llama-index-llms-ollama>=0.3.0
llama-index-embeddings-huggingface>=0.5.0
pydantic>=2.0.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
qdrant-client>=1.7.0
open-clip-torch>=2.20.0
pillow>=10.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
requests>=2.31.0
